Create separate aggregation handler
